Project refactor

+++ LabmdasSPC/src/main/java/cloud/lambdas/pojo/Company.java
@@ -1,20 +1,20 @@
 package cloud.lambdas.pojo;
 
 public class Company {
-    private Integer id;
+    private Long id;
     private String name;
 
     public Company() {}
-    public Company(Integer id, String name) {
+    public Company(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
